backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.1
pydantic[email]==2.7.1
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
pdfplumber==0.11.1
spacy==3.7.4
sentence-transformers==3.0.1
google-generativeai>=0.3.0
python-dotenv==1.0.1
alembic==1.13.1
langgraph==0.1.19
langchain-core==0.2.10
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
